- Fix api call tests - Adjust models to follow the inconsistent server responses

+++ PKOpenWeatherMapApi/Models/PKWindModel.swift
@@ -17,7 +17,7 @@ public class PKWindModel: PKJSONInitializable {
     }
     
     public let speed: Double
-    public let degree: Int
+    public let degree: Double
     public let gust: Int
     
     required init(with json:[String : Any]?) throws {
@@ -26,7 +26,7 @@ public class PKWindModel: PKJSONInitializable {
             throw PKSerializationError.missing(ObjectKeys.speed.rawValue, PKWindModel.self)
         }
 
-        guard let degree = json?[ObjectKeys.degree.rawValue] as? Int else {
+        guard let degree = json?[ObjectKeys.degree.rawValue] as? Double else {
             throw PKSerializationError.missing(ObjectKeys.degree.rawValue, PKWindModel.self)
         }
